+++ js/util.js
@@ -1,22 +1,3 @@
-const NAMES = [
-  'Иполит',
-  'Гаврила',
-  'Гурген',
-  'Инокентий',
-  'Аркадий',
-  'Валентин',
-  'Ашот',
-];
-
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
-]
-
 const isLengthFit = function (checkedLine, maxLength) {
   return checkedLine.length <= maxLength;
 }
@@ -46,8 +27,6 @@ const getRandomElement = function (array) {
 }
 
 export {
-  NAMES,
-  MESSAGES,
   isLengthFit,
   getRandomNumber,
   randomNumbersList,
